[Backport] 8065760: CHA: Improve abstract method support
Summary: CHA inline supports abstract method

Testing: test/hotspot/jtreg/compiler/cha/AbstractRootMethod.java

Reviewers: kuaiwei.kw, yulei, yude.lin

Issue: https://github.com/dragonwell-project/dragonwell11/issues/774
