+++ src/hotspot/share/code/dependencies.cpp
@@ -2236,10 +2236,12 @@ Method* Dependencies::find_unique_concrete_method(Klass* ctxk, Method* m, Klass*
   assert(fm == NULL || !fm->is_abstract(), "sanity");
   // Old CHA conservatively reports concrete methods in abstract classes
   // irrespective of whether they have concrete subclasses or not.
+  // Also, abstract root method case is not fully supported.
 #ifdef ASSERT
   Klass*  uniqp = NULL;
   Method* uniqm = Dependencies::find_unique_concrete_method(ctxk, m, &uniqp);
   assert(uniqm == NULL || uniqm == fm ||
+         m->is_abstract() ||
          uniqm->method_holder()->is_abstract() ||
          (fm == NULL && uniqm != NULL && uniqp != NULL && !InstanceKlass::cast(uniqp)->is_linked()),
          "sanity");
